+++ example/android/app/src/main/java/com/example/mobile_number_example/MainActivity.java
@@ -8,10 +8,10 @@ import io.flutter.embedding.engine.FlutterEngine;
 public class MainActivity extends FlutterActivity {
  
 
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    super.configureFlutterEngine(flutterEngine);
-    flutterEngine.getPlugins().add(new com.example.mobile_number.MobileNumberPlugin());
-  }
+  // @Override
+  // public void configureFlutterEngine(FlutterEngine flutterEngine) {
+  //   super.configureFlutterEngine(flutterEngine);
+  //   flutterEngine.getPlugins().add(new com.example.mobile_number.MobileNumberPlugin());
+  // }
 
 }