+++ example/lib/main.dart
@@ -1,9 +1,11 @@
 import 'dart:async';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:mobile_number/mobile_number.dart';
 import 'package:mobile_number/sim_card.dart';
+import 'package:mobile_number/widget_lifecycle.dart';
 
 void main() => runApp(MyApp());
 
@@ -19,11 +21,21 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
+    MobileNumber.listenPhonePermission((isPermissionGranted) {
+      if (isPermissionGranted) {
+        initMobileNumberState();
+      } else {}
+    });
+
     initMobileNumberState();
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initMobileNumberState() async {
+    if (!await MobileNumber.hasPhonePermission) {
+      await MobileNumber.requestPhonePermission;
+      return;
+    }
     String mobileNumber = '';
     // Platform messages may fail, so we use a try/catch PlatformException.
     try {
@@ -69,4 +81,6 @@ class _MyAppState extends State<MyApp> {
       ),
     );
   }
+
+  resumeCallback(void Function() param0) {}
 }