+++ lib/mobile_number.dart
@@ -1,10 +1,39 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
+import 'package:flutter/widgets.dart';
 import 'package:mobile_number/sim_card.dart';
+import 'package:mobile_number/widget_lifecycle.dart';
 
 class MobileNumber {
   static const MethodChannel _channel = const MethodChannel('mobile_number');
+  static const EventChannel _phonePermissionEvent =
+      EventChannel('phone_permission_event');
+
+//phone_permission_event // //
+
+  static void listenPhonePermission(
+      Function(bool isPermissionGranted) subscription) {
+    WidgetsBinding.instance.addObserver(WidgetLifecycle(
+      resumeCallBack: (() async {
+        if (await MobileNumber.hasPhonePermission) {
+          subscription(true);
+        } else {
+          subscription(false);
+        }
+      }),
+    ));
+  }
+
+  static Future<bool> get hasPhonePermission async {
+    final bool hasPermission =
+        await _channel.invokeMethod('hasPhonePermission');
+    return hasPermission;
+  }
+
+  static Future<void> get requestPhonePermission async {
+    await _channel.invokeMethod('requestPhonePermission');
+  }
 
   static Future<String> get mobileNumber async {
     final String simCardsJson = await _channel.invokeMethod('getMobileNumber');